Move code to individual files
